+++ src/connection/client.spec.ts
@@ -192,6 +192,7 @@ describe('DatabaseClient', () => {
   })
 
   describe('.insert()', () => {
+    type Song = typeof song & { id: number }
     const song = { name: 'Take On Me', artist: 'A-ha', rating: 5 }
 
     it('inserts the given record', async () => {
@@ -200,7 +201,9 @@ describe('DatabaseClient', () => {
       const { client } = mkClient()
       jest.spyOn(client, 'query').mockResolvedValue([newSong])
 
-      await expect(client.insert('song', song)).resolves.toBe(newSong)
+      // insert should be not nullable by default
+      const result: Song = await client.insert<Song>('song', song)
+      expect(result).toBe(newSong)
 
       expect(client.query).toHaveBeenCalledWith(
         expect.sqlMatching({
@@ -221,49 +224,22 @@ describe('DatabaseClient', () => {
       await expect(client.insert('song', song)).rejects.toThrow()
     })
 
+    it('returns null if no rows come back with onConflict=ignore', async () => {
+      const { client } = mkClient()
+      jest.spyOn(client, 'query').mockResolvedValue([])
+
+      // insert should be nullable when onConflict=ignore
+      const result: Song | null = await client.insert<Song>('song', song, {
+        onConflict: 'ignore',
+      })
+      expect(result).toBeNull()
+    })
+
     it('errors if multiple rows come back', async () => {
       const { client } = mkClient()
       jest.spyOn(client, 'query').mockResolvedValue([{}, {}])
 
       await expect(client.insert('song', song)).rejects.toThrow()
-    })
-  })
-
-  describe('.insertWith()', () => {
-    const song = { name: 'Take On Me', artist: 'A-ha', rating: 5 }
-
-    it('inserts the given record', async () => {
-      const newSong = { id: 1, ...song }
-
-      const { client } = mkClient()
-      jest.spyOn(client, 'query').mockResolvedValue([newSong])
-
-      await expect(client.insertWith('song', song, {})).resolves.toBe(newSong)
-
-      expect(client.query).toHaveBeenCalledWith(
-        expect.sqlMatching({
-          text: `
-            INSERT INTO "song" ("name","artist","rating")
-            VALUES ($1,$2,$3)
-            RETURNING *
-          `,
-          values: ['Take On Me', 'A-ha', 5],
-        }),
-      )
-    })
-
-    it('returns null if no rows come back', async () => {
-      const { client } = mkClient()
-      jest.spyOn(client, 'query').mockResolvedValue([])
-
-      await expect(client.insertWith('song', song, {})).resolves.toBeNull()
-    })
-
-    it('errors if multiple rows come back', async () => {
-      const { client } = mkClient()
-      jest.spyOn(client, 'query').mockResolvedValue([{}, {}])
-
-      await expect(client.insertWith('song', song, {})).rejects.toThrow()
     })
   })
 
@@ -275,7 +251,7 @@ describe('DatabaseClient', () => {
       ]
 
       const { client, mockQuery } = mkClient()
-      mockQuery.mockResolvedValue({ rows: [] })
+      mockQuery.mockResolvedValue({ rows: [{ id: 0 }] })
 
       await client.insertAll('song', songs)
 
@@ -308,7 +284,7 @@ describe('DatabaseClient', () => {
         if (query.values[0] === 'UNKNOWN') {
           throw new Error('fail')
         }
-        return { rows: [] }
+        return { rows: [{ id: 0 }] }
       })
 
       await expect(