+++ src/connection/insert.ts
@@ -4,11 +4,17 @@ export type InsertOptions = {
   onConflict?: ConflictOptions | null
 }
 
+export type InsertResult<T, Options extends InsertOptions> = ConflictType<
+  Options['onConflict']
+> extends 'ignore'
+  ? T | null
+  : T
+
 export const mkInsertQuery = <T extends Record<string, unknown>>(
   table: string,
   record: T,
   options: InsertOptions = {},
-) => {
+): SqlQuery => {
   const { onConflict = null } = options
 
   const columnNames = Object.keys(record)
@@ -27,6 +33,29 @@ export const mkInsertQuery = <T extends Record<string, unknown>>(
   `
 }
 
+export const toInsertResult = <T, Options extends InsertOptions>(
+  rows: T[],
+  options?: Options,
+): InsertResult<T, Options> => {
+  if (rows.length > 1) {
+    throw new Error(
+      `INSERT statement unexpectedly returned multiple rows: ${rows}`,
+    )
+  }
+
+  const row = rows.length === 1 ? rows[0] : null
+
+  const conflictOptions = options?.onConflict
+  const isNullableResult =
+    conflictOptions &&
+    (conflictOptions === 'ignore' || conflictOptions.action === 'ignore')
+  if (!isNullableResult && row === null) {
+    throw new Error(`INSERT statement unexpectedly returned no rows`)
+  }
+
+  return row as InsertResult<T, Options>
+}
+
 type ConflictTarget = { column: string } | { constraint: string }
 
 type ConflictOptions =
@@ -34,17 +63,25 @@ type ConflictOptions =
   | ({ action: 'ignore' } & Partial<ConflictTarget>)
   | ({ action: 'update' } & ConflictTarget)
 
+type ConflictType<Options> = Options extends ConflictOptions
+  ? Options extends 'ignore'
+    ? 'ignore'
+    : Options extends { action: 'ignore' }
+    ? 'ignore'
+    : 'update'
+  : null
+
 const mkConflictClause = (
   columnNamesSql: SqlQuery,
   valuesSql: SqlQuery,
   options: ConflictOptions | null,
-) => {
+): SqlQuery => {
   if (options === null) {
     return sql``
   }
 
   const { action, ...target } =
-    options === 'ignore' ? { action: 'ignore' } : options
+    options === 'ignore' ? { action: 'ignore' as const } : options
 
   const targetClause =
     'constraint' in target && target.constraint