+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Return inserted rows in `insertAll`
+* `insertAll` accepts partial records, just like `insert`
+* `insert` and `insertWith` are now combined into a single `insert` function. The function is smart enough to inspect the options and return `T | null` only if `onConflict=ignore`, and `T` otherwise.
 
 # 0.1.1